Move maf util to own artifact
