+++ maf/src/main/java/org/mskcc/cbio/maf/ValueTypeUtil.java
